+++ install.sh
@@ -23,8 +23,6 @@
 #                                                                                #
 ##################################################################################
 
-set -e
-
 # Check distrib
 if ! command -v apt-get &> /dev/null; then
   echo "ERROR: OS distribution not supported..."
@@ -57,29 +55,57 @@ echo ""
 
 echo "Adding i386 architecture..."
 dpkg --add-architecture i386 >/dev/null
+if [ "$?" -ne "0" ]; then
+  echo "ERROR: Cannot add i386 architecture..."
+  exit 1
+fi
 
 echo "Installing required packages..."
 apt-get update >/dev/null
 apt-get install -y -q libc6-i386 lib32stdc++6 lib32gcc1 lib32ncurses5 lib32z1 curl gdb screen tar wget >/dev/null
+if [ "$?" -ne "0" ]; then
+  echo "ERROR: Cannot install required packages..."
+  exit 1
+fi
 
-echo "Downloading CSGO Server Launcher..."
+echo "Downloading CSGO Server Launcher script..."
 wget ${baseUrl}/csgo-server-launcher.sh -O ${scriptPath} -q --no-check-certificate
+if [ "$?" -ne "0" ]; then
+  echo "ERROR: Cannot download CSGO Server Launcher script..."
+  exit 1
+fi
 
 echo "Chmod script..."
 chmod +x ${scriptPath}
+if [ "$?" -ne "0" ]; then
+  echo "ERROR: Cannot chmod CSGO Server Launcher script..."
+  exit 1
+fi
 
 echo "Install System-V style init script link..."
 update-rc.d csgo-server-launcher defaults >/dev/null
+if [ "$?" -ne "0" ]; then
+  echo "ERROR: Cannot install System-V style init script link..."
+  exit 1
+fi
 
 echo "Downloading CSGO Server Launcher configuration..."
 mkdir -p /etc/csgo-server-launcher/
 wget ${baseUrl}/csgo-server-launcher.conf -O ${confPath} -q --no-check-certificate
+if [ "$?" -ne "0" ]; then
+  echo "ERROR: Cannot download CSGO Server Launcher configuration..."
+  exit 1
+fi
 
 echo "Checking $user user exists..."
 getent passwd ${user} >/dev/null 2&>1
 if [ "$?" -ne "0" ]; then
   echo "Adding $user user..."
   useradd -m ${user}
+  if [ "$?" -ne "0" ]; then
+    echo "ERROR: Cannot add user $user..."
+    exit 1
+  fi
 else
   mkdir -p ~${user}
 fi